fix(ch. 21): provide crecise criteria on Ingredient.isSimple()
Provide a precise description of criteria of how to
determine the return value of Ingredient.isSimple(),
based on Lex's reply on discord. More information
may be found in the discussion of the original issue
ticket (i.e. GH-20).

TL;DR: isSimple() == true if and only if the criteria
is a combination of Item check and metadata check
(for those Item(s) that has hasSubtypes() == false,
the metadata is 0). Example of non-simple Ingredient
would be "true wildcard implementation" where the
wildcard value is completely ignored.
Remark: for regular ingredient, not specifying
metadata means "strictly match metadata to zero".

Close GH-20.