feat(ch. 11): introduce texture animation and convention on size
Per request in GH-21, we give a detailed overview on how to use
the animated texture mechanics to create more complex visual
effects.
This mechanics is often used in fluid model where the model is
static but the texture is animated so that people perceive
the model as "animated".
This still doesn't fully address GH-21 as it is talking about
the mechanics in general, and we need to set up the relation
between the fluid texture back and this part in order to fully
address GH-21.